// src/interfaces/rate-configurations.interface.ts
export interface RateConfigurationsInterface {
    id: string;
    name: string;
    is_shift_rate: boolean;
    program_id?: string;
    is_enabled?: boolean;
    is_deleted?: boolean;
    page?: string;
    limit?: string;
    created_by: string;
    updated_by: string;
    created_on?: bigint;
    updated_on?: bigint;
    hierarchies: any;
    job_templates: any;
    rate_configuration: any;
    job_type?: any;
    expenses:any;
}